Remove the model file from db interop
