+++ interop_query/views.py
@@ -2,8 +2,8 @@ from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_http_methods
 from django.db.models import Q
-from gene_function.models import GenomeInfo
-from interop_query.models import GeneAnnotations, GeneInfo
+from gene_function.models import GenomeInfo, GeneInfo
+from pangenome_analyses.models import GeneAnnotations
 
 import json
 import logging
